+++ scripts/graphiti_temporal_smoke/manifest.py
@@ -1,0 +1,175 @@
+"""External adapter manifest writer for the Graphiti/Zep smoke."""
+
+from __future__ import annotations
+
+from pathlib import Path
+from typing import Any
+
+from .common import rel, utc_now, write_json
+from .context import *  # noqa: F403
+from .models import StatusState
+
+def write_manifest(status: StatusState) -> dict[str, Any]:
+    """Write a generated external adapter manifest for this smoke."""
+
+    manifest = {
+        "schema": "elf.real_world_external_adapter_manifest/v1",
+        "manifest_id": f"graphiti-zep-temporal-smoke-{RUN_ID}",
+        "docker_isolation": {
+            "default": True,
+            "compose_file": "docker-compose.baseline.yml",
+            "runner": "scripts/graphiti-zep-docker-temporal-smoke.py",
+            "artifact_dir": "tmp/real-world-memory/graphiti-zep-smoke",
+            "host_global_installs_required": False,
+            "notes": [
+                f"Generated by the Graphiti/Zep Docker smoke at {utc_now()}.",
+                "The smoke uses generated public temporal facts and records typed setup/runtime failures.",
+            ],
+        },
+        "adapters": [
+            {
+                "adapter_id": "graphiti_zep_temporal_smoke",
+                "project": "Graphiti/Zep",
+                "adapter_kind": "docker_python_falkordb_temporal_smoke",
+                "evidence_class": status.evidence_class,
+                "docker_default": True,
+                "host_global_installs_required": False,
+                "overall_status": status.overall,
+                "setup": {
+                    "status": status.setup,
+                    "evidence": "The smoke runs inside the baseline Docker runner and uses Docker-local FalkorDB plus a container-local Python venv.",
+                    "command": "cargo make smoke-graphiti-zep-docker-temporal",
+                    "artifact": rel(OUT),
+                },
+                "run": {
+                    "status": status.run,
+                    "evidence": "The live path adds generated temporal fact triples and searches Graphiti/Zep for UUID, fact, valid_at, invalid_at, and source node evidence.",
+                    "command": "ELF_GRAPHITI_ZEP_SMOKE_START=1 ELF_GRAPHITI_ZEP_SMOKE_RUN=1 cargo make smoke-graphiti-zep-docker-temporal",
+                    "artifact": rel(OUT),
+                },
+                "result": {
+                    "status": status.result,
+                    "evidence": status.failure_reason
+                    if status.failure_reason
+                    else "Graphiti/Zep temporal search mapped current and historical facts to validity windows.",
+                    "artifact": rel(OUT),
+                },
+                "capabilities": [
+                    {
+                        "capability": "docker_falkordb_setup",
+                        "status": status.setup,
+                        "evidence": "The task starts a Docker Compose FalkorDB profile only when explicitly requested, and uses no host-global graph database.",
+                    },
+                    {
+                        "capability": "temporal_fact_triple_ingest",
+                        "status": status.run,
+                        "evidence": "The live worker uses Graphiti fact triples for current, historical, and rationale facts with validity windows.",
+                    },
+                    {
+                        "capability": "validity_window_evidence_mapping",
+                        "status": status.result,
+                        "evidence": "Search output UUID, fact text, valid_at, invalid_at, and node ids are mapped to memory_evolution expected evidence ids.",
+                    },
+                    {
+                        "capability": "quality_or_scale_claim",
+                        "status": "not_encoded",
+                        "evidence": "The smoke does not claim broad graph-memory quality, large-corpus behavior, managed Zep service behavior, or private-corpus performance.",
+                    },
+                ],
+                "suites": [
+                    {
+                        "suite_id": "memory_evolution",
+                        "status": status.result,
+                        "evidence": "Only generated current-versus-historical temporal relation facts are represented.",
+                    },
+                    {
+                        "suite_id": "retrieval",
+                        "status": status.run if status.run != "pass" else "not_encoded",
+                        "evidence": "Hybrid retrieval reachability is exercised by the live search, but broad retrieval quality scoring is not encoded.",
+                    },
+                    {
+                        "suite_id": "production_ops",
+                        "status": "not_encoded",
+                        "evidence": "The smoke records setup and provider boundaries but does not encode backup, restore, private corpus, or hosted-service operations.",
+                    },
+                ],
+                "scenarios": [
+                    {
+                        "scenario_id": "temporal_validity_window_mapping",
+                        "suite_id": "memory_evolution",
+                        "status": status.result,
+                        "elf_position": "untested",
+                        "comparison_outcome": "blocked"
+                        if status.result == "blocked"
+                        else "not_tested",
+                        "evidence": status.failure_reason
+                        if status.failure_reason
+                        else "Graphiti/Zep temporal search mapped generated current and historical relation facts to validity windows and evidence ids.",
+                        "command": "cargo make smoke-graphiti-zep-docker-temporal",
+                        "artifact": rel(OUT),
+                    }
+                ],
+                "evidence": [
+                    {"kind": "artifact", "ref": rel(OUT), "status": status.result},
+                    {"kind": "manifest", "ref": rel(MANIFEST_OUT), "status": status.overall},
+                    {"kind": "source", "ref": "https://github.com/getzep/graphiti", "status": "real"},
+                    {
+                        "kind": "source",
+                        "ref": "https://help.getzep.com/graphiti/getting-started/quick-start",
+                        "status": "real",
+                    },
+                    {
+                        "kind": "source",
+                        "ref": "https://help.getzep.com/graphiti/configuration/falkor-db-configuration",
+                        "status": "real",
+                    },
+                    {
+                        "kind": "source",
+                        "ref": "https://help.getzep.com/graphiti/working-with-data/adding-fact-triples",
+                        "status": "real",
+                    },
+                ],
+                "execution_metadata": {
+                    "sources": [
+                        {
+                            "label": "Graphiti repository",
+                            "url": "https://github.com/getzep/graphiti",
+                            "evidence": "Official source for the open-source temporal context graph engine.",
+                        },
+                        {
+                            "label": "Graphiti quick start",
+                            "url": "https://help.getzep.com/graphiti/getting-started/quick-start",
+                            "evidence": "Official search output examples include UUID, fact, valid_at, and invalid_at fields.",
+                        },
+                        {
+                            "label": "Graphiti FalkorDB configuration",
+                            "url": "https://help.getzep.com/graphiti/configuration/falkor-db-configuration",
+                            "evidence": "Official Docker-local FalkorDB setup and Python driver reference.",
+                        },
+                        {
+                            "label": "Graphiti fact triples",
+                            "url": "https://help.getzep.com/graphiti/working-with-data/adding-fact-triples",
+                            "evidence": "Official manual fact-triple ingest contract.",
+                        },
+                    ],
+                    "setup_path": "Run cargo make smoke-graphiti-zep-docker-temporal for a typed artifact; set ELF_GRAPHITI_ZEP_SMOKE_START=1 ELF_GRAPHITI_ZEP_SMOKE_RUN=1 with explicit provider configuration for a live attempt.",
+                    "runtime_boundary": "docker-compose.baseline.yml baseline-runner plus graphiti-zep FalkorDB profile, container-local Python venv, generated public temporal facts, and report artifacts under tmp/real-world-memory/graphiti-zep-smoke.",
+                    "resource_expectation": f"Graphiti package {GRAPHITI_REF}, fact_count=3, timeout_seconds={TIMEOUT_SECONDS}, FalkorDB host={FALKORDB_HOST}:{FALKORDB_PORT}.",
+                    "retry_guidance": [
+                        "Default command records a typed blocked artifact without model calls.",
+                        "Enable the live path only with Docker-local FalkorDB and explicit provider configuration.",
+                        "Treat missing validity windows or unmapped current/historical facts as wrong_result, not pass.",
+                    ],
+                    "research_depth": "D2 feasibility plus XY-888 Docker temporal smoke implementation; generated artifact decides live evidence class.",
+                },
+                "notes": [
+                    "The checked-in manifest record remains research_gate; generated smoke artifacts carry live status.",
+                    "Failure before Graphiti search output remains typed as blocked or incomplete.",
+                    "The smoke does not use a hosted Zep service, private corpora, or unrecorded provider credentials.",
+                ],
+            }
+        ],
+    }
+    write_json(MANIFEST_OUT, manifest)
+
+    return manifest
