+++ scripts/graphiti_temporal_smoke/summary.py
@@ -1,0 +1,35 @@
+"""Summary writer for the Graphiti/Zep smoke."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from .common import rel, utc_now, write_json
+from .context import MANIFEST_OUT, SUMMARY_OUT
+
+def write_summary(materialization: dict[str, Any], manifest: dict[str, Any], report: dict[str, Any]) -> None:
+    """Write a small summary artifact."""
+
+    write_json(
+        SUMMARY_OUT,
+        {
+            "schema": "elf.graphiti_zep_temporal_smoke_summary/v1",
+            "generated_at": utc_now(),
+            "adapter_id": "graphiti_zep_temporal_smoke",
+            "evidence_class": materialization["evidence_class"],
+            "status_boundary": {
+                "materialization": "setup/run/evidence-mapping state emitted by the smoke runner",
+                "manifest": "external adapter declaration consumed by the scorer",
+                "scored_benchmark": "post-score real_world_job outcome; use this for quality status",
+            },
+            "scored_benchmark": materialization["scored_benchmark"],
+            "materialization": materialization,
+            "manifest": {
+                "json": rel(MANIFEST_OUT),
+                "status_source": "external_adapter_manifest_pre_score",
+                "summary": manifest["adapters"][0]["overall_status"],
+                "suites": manifest["adapters"][0]["suites"],
+            },
+            "report": report,
+        },
+    )
