+++ scripts/ragflow_smoke/manifest.sh
@@ -1,0 +1,169 @@
+# RAGFlow Docker evidence smoke helper functions.
+# Sourced by scripts/ragflow-docker-evidence-smoke.sh.
+
+write_manifest() {
+	local generated_at out_rel manifest_rel retrieval_suite_status production_ops_status capability_retrieval_status capability_setup_status
+	generated_at="$(date -u +"%Y-%m-%dT%H:%M:%SZ")"
+	out_rel="$(relative_path "${OUT}")"
+	manifest_rel="$(relative_path "${MANIFEST_OUT}")"
+	retrieval_suite_status="$(json_status "${RESULT_STATUS}")"
+	capability_retrieval_status="$(json_status "${RESULT_STATUS}")"
+	capability_setup_status="$(json_status "${SETUP_STATUS}")"
+	production_ops_status="not_encoded"
+
+	jq -n \
+		--arg generated_at "${generated_at}" \
+		--arg manifest_id "ragflow-docker-evidence-smoke-${RUN_ID}" \
+		--arg out_rel "${out_rel}" \
+		--arg manifest_rel "${manifest_rel}" \
+		--arg evidence_class "${EVIDENCE_CLASS}" \
+		--arg overall_status "$(json_status "${OVERALL_STATUS}")" \
+		--arg setup_status "$(json_status "${SETUP_STATUS}")" \
+		--arg run_status "$(json_status "${RUN_STATUS}")" \
+		--arg result_status "$(json_status "${RESULT_STATUS}")" \
+		--arg retrieval_suite_status "${retrieval_suite_status}" \
+		--arg production_ops_status "${production_ops_status}" \
+		--arg capability_setup_status "${capability_setup_status}" \
+		--arg capability_retrieval_status "${capability_retrieval_status}" \
+		--arg ragflow_image "${RAGFLOW_IMAGE}" \
+		--arg cpu_gpu_mode "${CPU_GPU_MODE}" \
+		--arg failure_reason "${FAILURE_REASON}" \
+		--arg host_global_installs_required "${HOST_GLOBAL_INSTALLS_REQUIRED}" \
+		'{
+			schema: "elf.real_world_external_adapter_manifest/v1",
+			manifest_id: $manifest_id,
+			docker_isolation: {
+				default: true,
+				compose_file: "official RAGFlow docker/docker-compose.yml",
+				runner: "scripts/ragflow-docker-evidence-smoke.sh",
+				artifact_dir: "tmp/real-world-memory/ragflow-smoke",
+				host_global_installs_required: ($host_global_installs_required == "true"),
+				notes: [
+					"Generated by the RAGFlow evidence-smoke script at " + $generated_at + ".",
+					"The smoke uses a generated public corpus and does not use private corpus or operator-owned provider credentials."
+				]
+			},
+			adapters: [
+				{
+					adapter_id: "ragflow_docker_evidence_smoke",
+					project: "RAGFlow",
+					adapter_kind: "docker_service_evidence_smoke",
+					evidence_class: $evidence_class,
+					docker_default: true,
+					host_global_installs_required: ($host_global_installs_required == "true"),
+					overall_status: $overall_status,
+					setup: {
+						status: $setup_status,
+						evidence: "Official RAGFlow Docker Compose boundary and resource envelope were evaluated for the tiny evidence smoke.",
+						command: "cargo make smoke-ragflow-docker",
+						artifact: $out_rel
+					},
+					run: {
+						status: $run_status,
+						evidence: "The smoke attempts dataset creation, empty-document corpus ingest, chunk insert, retrieval query, and reference chunk extraction.",
+						command: "ELF_RAGFLOW_SMOKE_START=1 ELF_RAGFLOW_SMOKE_ACCEPT_RESOURCE_ENVELOPE=1 cargo make smoke-ragflow-docker",
+						artifact: $out_rel
+					},
+					result: {
+						status: $result_status,
+						evidence: (
+							if $failure_reason == "" then "Returned RAGFlow reference chunks were mapped to generated real_world_job evidence ids for the smoke only."
+							else $failure_reason
+							end
+						),
+						artifact: $out_rel
+					},
+					capabilities: [
+						{
+							capability: "official_docker_service_boundary",
+							status: $capability_setup_status,
+							evidence: "The script uses the official RAGFlow Docker Compose setup and records image, disk, startup, CPU/GPU, and vm.max_map_count evidence."
+						},
+						{
+							capability: "dataset_or_chunk_ingest",
+							status: $run_status,
+							evidence: "The live path creates a generated public dataset, empty document, and chunk before querying."
+						},
+						{
+							capability: "retrieval_reference_mapping",
+							status: $capability_retrieval_status,
+							evidence: "The script maps returned chunk id, document id, document name, dataset id, positions, and similarity fields to benchmark evidence ids."
+						},
+						{
+							capability: "quality_or_scale_claim",
+							status: "not_encoded",
+							evidence: "The smoke does not run broad RAGFlow quality scoring, scale tests, private corpora, or comparative ranking claims."
+						}
+					],
+					suites: [
+						{
+							suite_id: "retrieval",
+							status: $retrieval_suite_status,
+							evidence: "Only the generated-public RAGFlow evidence-smoke retrieval path is represented."
+						},
+						{
+							suite_id: "production_ops",
+							status: $production_ops_status,
+							evidence: "Resource envelope evidence is recorded, but no production-ops suite scoring is encoded."
+						},
+						{
+							suite_id: "knowledge_compilation",
+							status: "not_encoded",
+							evidence: "RAGFlow page or knowledge-compilation behavior is not part of this smoke."
+						}
+					],
+					evidence: [
+						{
+							kind: "artifact",
+							ref: $out_rel,
+							status: $result_status
+						},
+						{
+							kind: "manifest",
+							ref: $manifest_rel,
+							status: $overall_status
+						},
+						{
+							kind: "source",
+							ref: "https://ragflow.io/docs/",
+							status: "real"
+						},
+						{
+							kind: "source",
+							ref: "https://raw.githubusercontent.com/infiniflow/ragflow/main/docs/references/http_api_reference.md",
+							status: "real"
+						}
+					],
+					execution_metadata: {
+						sources: [
+							{
+								label: "RAGFlow quickstart",
+								url: "https://ragflow.io/docs/",
+								evidence: "Official Docker startup, resource envelope, vm.max_map_count, and provider configuration guidance."
+							},
+							{
+								label: "RAGFlow HTTP API reference",
+								url: "https://raw.githubusercontent.com/infiniflow/ragflow/main/docs/references/http_api_reference.md",
+								evidence: "Official dataset, document, chunk, retrieval, and reference-chunk field contract."
+							}
+						],
+						setup_path: "Run the official RAGFlow Docker Compose stack with generated public corpus only.",
+						runtime_boundary: "Official RAGFlow Docker Compose service boundary; no host-global RAGFlow install.",
+						resource_expectation: (
+							"RAGFlow image " + $ragflow_image + ", CPU/GPU mode " + $cpu_gpu_mode + ", official minimums 4 CPU cores, 16 GB RAM, 50 GB disk, and vm.max_map_count >= 262144."
+						),
+						retry_guidance: [
+							"Default command records a typed blocked preflight unless resource-heavy startup is explicitly enabled.",
+							"Set ELF_RAGFLOW_SMOKE_START=1 and ELF_RAGFLOW_SMOKE_ACCEPT_RESOURCE_ENVELOPE=1 for a live Docker startup attempt.",
+							"Provide only a local self-hosted RAGFlow API key; do not use private corpora or operator-owned model provider credentials for this smoke."
+						],
+						research_depth: "D2 feasibility plus XY-885 evidence-smoke implementation; generated artifact decides live evidence class."
+					},
+					notes: [
+						"This adapter record is generated by a smoke artifact and must not be generalized into broad RAGFlow quality evidence.",
+						"Failure before query output remains typed as blocked, incomplete, or not_encoded."
+					]
+				}
+			]
+		}' >"${MANIFEST_OUT}"
+}
