+++ scripts/ragflow_smoke/scoring.sh
@@ -1,0 +1,60 @@
+# RAGFlow Docker evidence smoke helper functions.
+# Sourced by scripts/ragflow-docker-evidence-smoke.sh.
+
+write_scored_benchmark() {
+	if [[ -s "${REPORT_JSON}" ]]; then
+		jq 'def count($key): (.summary[$key] // 0);
+			def scored_status:
+				if count("wrong_result") > 0 then "wrong_result"
+				elif count("lifecycle_fail") > 0 then "lifecycle_fail"
+				elif count("incomplete") > 0 then "incomplete"
+				elif count("blocked") > 0 then "blocked"
+				elif count("not_encoded") > 0 then "not_encoded"
+				elif count("pass") > 0 then "pass"
+				else "not_encoded"
+				end;
+			{
+				schema: "elf.scored_benchmark_status/v1",
+				source: "real_world_job_benchmark",
+				status: scored_status,
+				counts: {
+					pass: count("pass"),
+					wrong_result: count("wrong_result"),
+					lifecycle_fail: count("lifecycle_fail"),
+					incomplete: count("incomplete"),
+					blocked: count("blocked"),
+					not_encoded: count("not_encoded")
+				},
+				job_count: (.summary.job_count // 0),
+				mean_score: (.summary.mean_score // null),
+				evidence_coverage: (.summary.evidence_coverage // null)
+			}' "${REPORT_JSON}" >"${SCORED_BENCHMARK}"
+	else
+		jq -n '{
+			schema: "elf.scored_benchmark_status/v1",
+			source: "real_world_job_benchmark",
+			status: "pending",
+			reason: "The smoke materialization was written before benchmark scoring completed."
+		}' >"${SCORED_BENCHMARK}"
+	fi
+}
+
+write_scored_report() {
+	(
+		cd "${ROOT_DIR}"
+		cargo run -p elf-eval --bin real_world_job_benchmark -- run \
+			--fixtures "${FIXTURE_PATH}" \
+			--out "${REPORT_JSON}" \
+			--run-id real-world-memory-live-ragflow \
+			--adapter-id ragflow_docker_evidence_smoke \
+			--adapter-name "RAGFlow Docker evidence smoke adapter" \
+			--adapter-behavior docker_service_evidence_smoke \
+			--adapter-storage-status "$(json_status "${SETUP_STATUS}")" \
+			--adapter-runtime-status "$(json_status "${OVERALL_STATUS}")" \
+			--adapter-notes "Generated by the RAGFlow Docker evidence smoke; pass or wrong_result requires reference chunks mapped to generated evidence ids, while resource/setup/API-key limits remain typed." \
+			--external-adapter-manifest "${MANIFEST_OUT}"
+		cargo run -p elf-eval --bin real_world_job_benchmark -- publish \
+			--report "${REPORT_JSON}" \
+			--out "${REPORT_MD}"
+	)
+}
